% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_box.R
\name{boxplot_residuals}
\alias{boxplot_residuals}
\title{Boxplot of Residuals}
\usage{
boxplot_residuals(residuals)
}
\arguments{
\item{residuals}{A numeric vector of residuals (difference between actual and predicted values)}
}
\description{
This function creates a boxplot displaying the distribution of residuals.
}
\examples{
boston_data <- load_boston_data("path")
preprocessed_data <- preprocess_boston_data(boston_data)
actual_values <- preprocessed_data$test_data$medv
predicted_values <- predictions$pred_medv
residuals <- actual_values - predicted_values
boxplot_residuals(residuals)
}
