% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_squared.R
\name{calculate_r_squared}
\alias{calculate_r_squared}
\title{Calculate the R-squared Value of a Linear Regression Model}
\usage{
calculate_r_squared(model, test_data)
}
\arguments{
\item{model}{A linear regression model object created using lm()}

\item{test_data}{A dataframe containing the test dataset}
}
\value{
A single numeric value representing the R-squared value of the model
}
\description{
This function calculates and returns the R-squared value of a given linear regression model.
}
\examples{
boston_data <- load_boston_data("path")
preprocessed_data <- preprocess_boston_data(boston_data)
train_result <- train_linear_regression(preprocessed_data$train_data)
r_squared_value <- calculate_r_squared(train_result$model, preprocessed_data$test_data)
}
