% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_heatmap.R
\name{heatmap_correlation}
\alias{heatmap_correlation}
\title{Heatmap of Correlation between Features}
\usage{
heatmap_correlation(dataset)
}
\arguments{
\item{dataset}{A dataframe containing the dataset with features}
}
\description{
This function creates a heatmap displaying the correlation between features in the given dataset.
}
\examples{
boston_data <- load_boston_data("path")
preprocessed_data <- preprocess_boston_data(boston_data)
heatmap_correlation(preprocessed_data$train_data)
}
