% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_boston_data}
\alias{load_boston_data}
\title{Load Boston Housing dataset}
\usage{
load_boston_data(package_path)
}
\arguments{
\item{package_path}{The path to the package containing the Boston Housing dataset}
}
\value{
A list containing two data.frames: train_data and test_data
}
\description{
This function loads the Boston Housing dataset saved as a CSV file in the "data" directory.
}
\examples{
boston_data <- load_boston_data("path")
}
