% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{make_predictions}
\alias{make_predictions}
\title{Make predictions using the trained linear regression model}
\usage{
make_predictions(test_data, model)
}
\arguments{
\item{test_data}{A data.frame containing the testing set of the preprocessed Boston Housing dataset}

\item{model}{A list with the trained linear regression model and the training data}
}
\value{
A data.frame with the predicted values and actual values
}
\description{
This function makes predictions using the trained linear regression model and the testing set of the Boston Housing dataset.
}
\examples{
boston_data <- load_boston_data("path")
preprocessed_data <- preprocess_boston_data(boston_data)
train_result <- train_linear_regression(preprocessed_data$train_data)
predictions <- make_predictions(preprocessed_data$test_data, train_result)
}
