% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_scatter.R
\name{scatter_actual_vs_predicted}
\alias{scatter_actual_vs_predicted}
\title{Scatter plot of Actual vs. Predicted Values}
\usage{
scatter_actual_vs_predicted(actual_values, predicted_values)
}
\arguments{
\item{actual_values}{A numeric vector of actual target values}

\item{predicted_values}{A numeric vector of predicted target values}
}
\description{
This function creates a scatter plot comparing actual and predicted values of the target variable.
}
\examples{
boston_data <- load_boston_data("path")
preprocessed_data <- preprocess_boston_data(boston_data)
train_result <- train_linear_regression(preprocessed_data$train_data)
predictions <- make_predictions(preprocessed_data$test_data, train_result)
actual_values <- preprocessed_data$test_data$medv
predicted_values <- predictions$pred_medv
scatter_actual_vs_predicted(actual_values, predicted_values)
}
